Added Authentication Controllers and functionality (Login , Register , Logout)

+++ Server/controllers/tasks.py
@@ -6,12 +6,13 @@ from sqlalchemy.orm import Session
 from fastapi import Depends
 from fastapi.responses import JSONResponse
 from schema.tasks_shemas import taskSchema, UpdateTaskSchema
-models.tasks_model.Base.metadata.create_all(engine)
 from utils.exception import * 
 from utils.response import CustomResponse
+from Middleware.auth_middleware import verify_jwt
+from fastapi import HTTPException,status
 
 ## Fucntion to create a Task 
-def createTask( task : taskSchema , db: Session = Depends(ConnectDB) ):
+def createTask( task : taskSchema , db: Session = Depends(ConnectDB), current_user = Depends(verify_jwt) ):
     '''
     This function creates a task in tha database.
     '''
@@ -22,7 +23,8 @@ def createTask( task : taskSchema , db: Session = Depends(ConnectDB) ):
         name=task.name,
         description=task.description,
         isDone=task.isDone,
-        date=task.date
+        date=task.date,
+        createdBy = current_user['id']
     )
         
         db.add(new_task)
@@ -37,12 +39,16 @@ def createTask( task : taskSchema , db: Session = Depends(ConnectDB) ):
     
     
 ## Fucntion to get all the tasks from the database
-def getAllTasks(db:Session = Depends(ConnectDB)):
+def getAllTasks(db:Session = Depends(ConnectDB), current_user = Depends(verify_jwt)):
     
     '''
     This function fetches all the tasks from the database.
     '''
-    tasks = db.query(TaskModel).all()
+    user_id = current_user["id"]
+    print('UserId : ' , user_id)
+
+    tasks = db.query(TaskModel).filter(TaskModel.createdBy == user_id).all()
+    print('Task : ' , tasks)
     if tasks :
          return CustomResponse.success(
            message='All Tasks Fetched Successfully!',
@@ -52,9 +58,14 @@ def getAllTasks(db:Session = Depends(ConnectDB)):
         return TaskNotFound(name = f'No Tasks')
 
 
-def getTaskById(id: int, db: Session = Depends(ConnectDB)):
+def getTaskById(id: int, db: Session = Depends(ConnectDB), current_user = Depends(verify_jwt)):
     task = db.query(TaskModel).filter(TaskModel.id == id).first()
     if task:
+        if task.createdBy != current_user["id"]:
+            raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to access this task"
+        )
         return CustomResponse.success(
            message='Task Fetched Successfully!',
            data= jsonable_encoder(task)
@@ -64,7 +75,7 @@ def getTaskById(id: int, db: Session = Depends(ConnectDB)):
 
 
 
-def deleteTaskById(id:int,db: Session = Depends(ConnectDB)):
+def deleteTaskById(id:int,db: Session = Depends(ConnectDB), current_user = Depends(verify_jwt)):
     
     '''
     This function Fetches a single task by the given id 
@@ -72,6 +83,12 @@ def deleteTaskById(id:int,db: Session = Depends(ConnectDB)):
     task = db.query(TaskModel).filter(TaskModel.id == id).first()
     
     if task:
+        if task.createdBy != current_user["id"]:
+            raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to access this task"
+        )
+
         db.delete(task)
         db.commit()
         return CustomResponse.success(
@@ -82,12 +99,17 @@ def deleteTaskById(id:int,db: Session = Depends(ConnectDB)):
          return TaskNotFound(name=f'id : {id}')
 
 
-def updateTaskId(updated_task : UpdateTaskSchema , id:int ,db: Session = Depends(ConnectDB)):
+def updateTaskId(updated_task : UpdateTaskSchema , id:int ,db: Session = Depends(ConnectDB), current_user = Depends(verify_jwt)):
     '''
     This function updates the task with the help of given id 
     '''
     task = db.query(TaskModel).filter(TaskModel.id == id).first()
     if task:
+        if task.createdBy != current_user["id"]:
+            raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to access this task"
+        )
         update_data = updated_task.model_dump(exclude_unset=True)
         for key , value in update_data.items():
             setattr(task, key , value)
@@ -102,7 +124,7 @@ def updateTaskId(updated_task : UpdateTaskSchema , id:int ,db: Session = Depends
     
 
 
-def toggleIsDoneFlag(id: int, db: Session = Depends(ConnectDB)):
+def toggleIsDoneFlag(id: int, db: Session = Depends(ConnectDB), current_user = Depends(verify_jwt)):
     '''
     This function  toggles the status of the task from done to
     undone and vice versa
@@ -110,6 +132,11 @@ def toggleIsDoneFlag(id: int, db: Session = Depends(ConnectDB)):
     task = db.query(TaskModel).filter(TaskModel.id == id).first()
     
     if task:
+        if task.createdBy != current_user["id"]:
+            raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to access this task"
+        )
         task.isDone = not task.isDone  
         db.commit()                    
         db.refresh(task)               #
